# docs/swagger.yaml
definitions:
...
  common.Currency:
    enum:
    - 101
    - 102
    - 103
    - 104
    - 105
    - 106
    - 107
    - 108
    - 109
    - 110
    - 111
    - 112
    - 113
    - 114
    - 115
    - 116
    - 117
    - 118
    - 201
    - 202
    - 203
    - 204
    - 205
    - 206
    - 207
    - 208
    - 209
    - 210
    - 211
    - 212
    - 213
    - 214
    - 215
    - 216
    - 217
    - 218
    - 219
    - 220
    - 221
    - 222
    - 223
    - 224
    - 225
    - 226
    - 227
    - 228
    - 229
    - 230
    - 231
    - 232
    - 233
    - 234
    - 235
    - 236
    - 237
    - 238
    - 239
    - 240
    - 299
    - 1001
    type: integer
    x-enum-varnames:
    - CURRENCY_USDT
    - CURRENCY_ETH
    - CURRENCY_BTC
    - CURRENCY_USDC
    - CURRENCY_DAI
    - CURRENCY_WBTC
    - CURRENCY_TRX
    - CURRENCY_ADA
    - CURRENCY_BCH
    - CURRENCY_DOGE
    - CURRENCY_LTC
    - CURRENCY_XRP
    - CURRENCY_SOL
    - CURRENCY_BNB
    - CURRENCY_ETC
    - CURRENCY_MATIC
    - CURRENCY_TON
    - CURRENCY_AVAX
    - CURRENCY_USD
    - CURRENCY_EUR
    - CURRENCY_JPY
    - CURRENCY_GBP
    - CURRENCY_CNY
    - CURRENCY_CAD
    - CURRENCY_AUD
    - CURRENCY_CHF
    - CURRENCY_HKD
    - CURRENCY_SEK
    - CURRENCY_NZD
    - CURRENCY_MXN
    - CURRENCY_SGD
    - CURRENCY_NOK
    - CURRENCY_KRW
    - CURRENCY_TRY
    - CURRENCY_INR
    - CURRENCY_BRL
    - CURRENCY_ZAR
    - CURRENCY_RUB
    - CURRENCY_TWD
    - CURRENCY_DKK
    - CURRENCY_PLN
    - CURRENCY_ILS
    - CURRENCY_AED
    - CURRENCY_ARS
    - CURRENCY_MYR
    - CURRENCY_THB
    - CURRENCY_SAR
    - CURRENCY_CZK
    - CURRENCY_UAH
    - CURRENCY_PHP
    - CURRENCY_MOP
    - CURRENCY_EGP
    - CURRENCY_MAD
    - CURRENCY_KES
    - CURRENCY_RON
    - CURRENCY_BGN
    - CURRENCY_VND
    - CURRENCY_COP
    - CURRENCY_OTHER_FIAT
    - CURRENCY_EPOINT
  common.CurrencyType:
    enum:
    - 1
    - 2
    - 4
    type: integer
    x-enum-varnames:
    - CURRENCY_TYPE_CRYPTO
    - CURRENCY_TYPE_FIAT
    - CURRENCY_TYPE_POINT
...
  common.RatePurpose:
    enum:
    - 1
    - 2
    - 3
    - 4
    - 5
    - 6
    - 7
    - 8
    - 9
    - 10
    - 11
    - 12
    - 13
    - 14
    type: integer
    x-enum-varnames:
    - RATE_PURPOSE_APPLY
    - RATE_PURPOSE_EXCHANGE
    - RATE_PURPOSE_TRANSFER
    - RATE_PURPOSE_CARD_TO_CARD
    - RATE_PURPOSE_TOP_UP
    - RATE_PURPOSE_TOP_DOWN
    - RATE_PURPOSE_FEE
    - RATE_PURPOSE_LIMIT
    - RATE_PURPOSE_MERCHANT_EXCHANGE
    - RATE_PURPOSE_MERCHANT_TRANSFER
    - RATE_PURPOSE_DISCOUNT
    - RATE_PURPOSE_DISPLAY
    - RATE_PURPOSE_CHIPPAY_EXPRESS_BUY
    - RATE_PURPOSE_TOP_UP_POINT
...
  shared-modules_entities.CurrencyVO:
    properties:
      code:
        $ref: '#/definitions/common.Currency'
      name:
        type: string
      type:
        $ref: '#/definitions/common.CurrencyType'
    type: object
...
  shared-modules_entities.ExchangePreviewVO:
    properties:
      displayRate:
        $ref: '#/definitions/shared-modules_entities.ExchangeRateVO'
      exchangeFee:
        type: string
      expiredAt:
        example: "0"
        type: string
      fromAmount:
        type: string
      fromCardID:
        example: "0"
        type: string
      fromCategory:
        type: string
      fromCategoryID:
        example: "0"
        type: string
      fromCurrency:
        type: string
      key:
        type: string
      rate:
        items:
          $ref: '#/definitions/shared-modules_entities.ExchangeRateVO'
        type: array
      toAmount:
        type: string
      toCardID:
        example: "0"
        type: string
      toCategory:
        type: string
      toCategoryID:
        example: "0"
        type: string
      toCurrency:
        type: string
    type: object
  shared-modules_entities.ExchangeRateVO:
    properties:
      baseCurrency:
        type: string
      purpose:
        type: string
      quoteCurrency:
        type: string
      rate:
        type: number
      timestamp:
        example: "0"
        type: string
    type: object
...
  shared-modules_entities.GetExchangeRateForm:
    properties:
      baseCurrencies:
        items:
          type: string
        type: array
      purpose:
        $ref: '#/definitions/common.RatePurpose'
      quoteCurrencies:
        items:
          type: string
        type: array
    required:
    - baseCurrencies
    - quoteCurrencies
    type: object
...
  shared-modules_entities.ListExchangeRateForm:
    properties:
      baseCurrency:
        type: string
      purpose:
        $ref: '#/definitions/common.RatePurpose'
      quoteCurrencies:
        items:
          type: string
        type: array
    required:
    - baseCurrency
    - quoteCurrencies
    type: object
  shared-modules_entities.ListExchangeRateVO:
    properties:
      effectiveRates:
        additionalProperties:
          items:
            $ref: '#/definitions/shared-modules_entities.ExchangeRateVO'
          type: array
        type: object
      fees:
        additionalProperties:
          type: number
        type: object
      records:
        items:
          $ref: '#/definitions/shared-modules_entities.ExchangeRateVO'
        type: array
    type: object
...
  shared-modules_entities.ListSystemParametersVO:
    properties:
      records:
        items:
          $ref: '#/definitions/shared-modules_entities.SystemParameterVO'
        type: array
    type: object
...
  shared-modules_entities.SystemParameterVO:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        example: "0"
        type: string
      name:
        type: string
      updated_at:
        type: string
      value:
        type: string
    type: object
...
  shared-modules_entities.TransferConfirmForm:
    properties:
      deviceId:
        type: string
      id:
        type: string
      pinCode:
        type: string
      rawId:
        type: string
      response:
        properties:
          authenticatorData:
            type: string
          clientDataJSON:
            type: string
          signature:
            type: string
          userHandle:
            type: string
        type: object
      transferKey:
        type: string
      type:
        type: string
    required:
    - transferKey
    type: object
  shared-modules_entities.TransferConfirmVO:
    properties:
      orderNO:
        type: string
    type: object
  shared-modules_entities.TransferPreviewForm:
    properties:
      fromAmount:
        example: "0"
        type: string
      fromCardId:
        example: "0"
        type: string
      mainnet:
        type: string
      protocol:
        type: string
      remark:
        type: string
      toAddress:
        type: string
      toAmount:
        example: "0"
        type: string
      toCardId:
        example: "0"
        type: string
      toEmail:
        type: string
      toUserId:
        example: "0"
        type: string
    required:
    - fromCardId
    type: object
...
info:
  contact: {}
paths:
...
  /web/quote/exchangeRates/getRates:
    post:
      description: List real-time exchange rate data.
      parameters:
      - description: body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/shared-modules_entities.GetExchangeRateForm'
      - description: User token
        in: header
        name: X-Token
        required: true
        type: string
      - description: accept language
        in: header
        name: Accept-Language
        type: string
      - description: Extend
        in: header
        name: X-Extend
        type: string
      - description: Convert
        in: header
        name: X-Convert
        type: string
      responses:
        "0":
          description: data
          schema:
            $ref: '#/definitions/shared-modules_entities.ListExchangeRateVO'
      tags:
      - web/quote
  /web/quote/exchangeRates/list:
    post:
      description: List real-time exchange rate data.
      parameters:
      - description: body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/shared-modules_entities.ListExchangeRateForm'
      - description: User token
        in: header
        name: X-Token
        required: true
        type: string
      - description: accept language
        in: header
        name: Accept-Language
        type: string
      - description: Extend
        in: header
        name: X-Extend
        type: string
      - description: Convert
        in: header
        name: X-Convert
        type: string
      responses:
        "0":
          description: data
          schema:
            $ref: '#/definitions/shared-modules_entities.ListExchangeRateVO'
      tags:
      - web/quote
  /web/system/config/list:
    post:
      description: List all system settings.
      parameters:
      - description: User token
        in: header
        name: X-Token
        required: true
        type: string
      - description: accept language
        in: header
        name: Accept-Language
        type: string
      - description: Extend
        in: header
        name: X-Extend
        type: string
      - description: Convert
        in: header
        name: X-Convert
        type: string
      responses:
        "0":
          description: data
          schema:
            $ref: '#/definitions/shared-modules_entities.ListSystemParametersVO'
      tags:
      - web/system
  /web/system/currencies:
    post:
      description: List all currencies.
      parameters:
      - description: User token
        in: header
        name: X-Token
        required: true
        type: string
      - description: accept language
        in: header
        name: Accept-Language
        type: string
      - description: Extend
        in: header
        name: X-Extend
        type: string
      - description: Convert
        in: header
        name: X-Convert
        type: string
      responses:
        "0":
          description: data
          schema:
            $ref: '#/definitions/shared-modules_entities.CurrencyVO'
      tags:
      - web/system
  /web/system/parameters:
    post:
      description: List all system parameters.
      parameters:
      - description: User token
        in: header
        name: X-Token
        required: true
        type: string
      - description: accept language
        in: header
        name: Accept-Language
        type: string
      - description: Extend
        in: header
        name: X-Extend
        type: string
      - description: Convert
        in: header
        name: X-Convert
        type: string
      responses:
        "0":
          description: data
          schema:
            $ref: '#/definitions/shared-modules_entities.ListSystemParametersVO'
      tags:
      - web/system
...
  /web/transfer/confirm:
    post:
      description: Apply for a new transfer confirmation.
      parameters:
      - description: body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/shared-modules_entities.TransferConfirmForm'
      - description: User token
        in: header
        name: X-Token
        required: true
        type: string
      - description: accept language
        in: header
        name: Accept-Language
        type: string
      - description: Extend
        in: header
        name: X-Extend
        type: string
      - description: Convert
        in: header
        name: X-Convert
        type: string
      responses:
        "0":
          description: data
          schema:
            $ref: '#/definitions/shared-modules_entities.TransferConfirmVO'
      summary: Apply for a new transfer confirmation. 轉帳(send發送) 自己的數幣轉別人的數幣
      tags:
      - web/transfer
  /web/transfer/preview:
    post:
      description: Apply for a new transfer preview.
      parameters:
      - description: body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/shared-modules_entities.TransferPreviewForm'
      - description: User token
        in: header
        name: X-Token
        required: true
        type: string
      - description: accept language
        in: header
        name: Accept-Language
        type: string
      - description: Extend
        in: header
        name: X-Extend
        type: string
      - description: Convert
        in: header
        name: X-Convert
        type: string
      responses:
        "0":
          description: data
          schema:
            $ref: '#/definitions/shared-modules_entities.ExchangePreviewVO'
      summary: apply for a new transfer preview. 轉帳(send發送) 自己的數幣轉別人的數幣
      tags:
      - web/transfer
...
swagger: "2.0"